Detect and fix broken src git repo association
The container's src directory is a git repository with a relationship
to another git repository (in ~/conf/git/ikiwiki). If ~/conf is volume
mounted, the relationship between the two repositories can break
(because the ref for HEAD in the src directory does not exist in the
volume-mounted backing git repo).

Detect this (by checking the exit status of "git show-ref", which fails
in this situation) and re-clone the repository on container start-up if
necessary.

We might need to do the same for "libdir" and "templates", too: they are
set up a different way (not --shared, yet; no commits in the cloned
repos) and might need a different approach to fix.